refactor: 💡 Move Components to components dir
